feat(api): move authorization constants in shared context
