+++ api/src/identity-access-management/domain/usecases/anonymize-gar-authentication-methods.usecase.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 
-import { PIX_ADMIN } from '../../../authorization/domain/constants.js';
+import { PIX_ADMIN } from '../../../shared/domain/constants.js';
 import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 import { AuditLoggingJob } from '../../../shared/domain/models/jobs/AuditLoggingJob.js';
