+++ api/src/identity-access-management/domain/usecases/authenticate-user.usecase.js
@@ -1,5 +1,5 @@
-import { PIX_ADMIN } from '../../../authorization/domain/constants.js';
 import { config } from '../../../shared/config.js';
+import { PIX_ADMIN } from '../../../shared/domain/constants.js';
 import { ForbiddenAccess, UserNotFoundError } from '../../../shared/domain/errors.js';
 import { NON_OIDC_IDENTITY_PROVIDERS } from '../constants/identity-providers.js';
 import { createWarningConnectionEmail } from '../emails/create-warning-connection.email.js';
