+++ api/src/prescription/campaign/dependencies.json
@@ -1,7 +1,6 @@
 {
   "name": "prescription/campaign",
   "dependsOn": [
-    "authorization",
     "devcomp",
     "evaluation",
     "identity-access-management",
