+++ api/src/shared/domain/constants.js
@@ -183,10 +183,33 @@ const constants = {
   STUDENT_RECONCILIATION_ERRORS,
 };
 
+const PIX_ADMIN = {
+  NOT_ALLOWED_MSG: "Vous n'avez pas les droits pour vous connecter.",
+  ROLES: {
+    CERTIF: 'CERTIF',
+    METIER: 'METIER',
+    SUPER_ADMIN: 'SUPER_ADMIN',
+    SUPPORT: 'SUPPORT',
+  },
+};
+
+const PIX_ORGA = {
+  ROLES: {
+    ADMIN: 'ORGA_ADMIN',
+  },
+};
+
+const CLIENTS = {
+  SCRIPT: 'SCRIPT',
+  ADMIN: 'PIX_ADMIN',
+  ORGA: 'PIX_ORGA',
+};
+
 export {
   ALL_TREATMENTS,
   CAMPAIGN_FEATURES,
   CERTIFICATION_CENTER_TYPES,
+  CLIENTS,
   COMPETENCES_COUNT,
   constants,
   DEFAULT_LEVEL_FOR_FIRST_CHALLENGE,
@@ -207,7 +230,9 @@ export {
   MINIMUM_REPRODUCIBILITY_RATE_TO_BE_TRUSTED,
   OIDC_ERRORS,
   ORGANIZATION_FEATURE,
+  PIX_ADMIN,
   PIX_COUNT_BY_LEVEL,
+  PIX_ORGA,
   PIX_ORIGIN,
   STUDENT_RECONCILIATION_ERRORS,
   V3_REPRODUCIBILITY_RATE,
