+++ api/tests/certification/evaluation/acceptance/scoring-and-capacity-simulator-route_test.js
@@ -1,5 +1,5 @@
 import { createServer } from '../../../../server.js';
-import { PIX_ADMIN } from '../../../../src/authorization/domain/constants.js';
+import { PIX_ADMIN } from '../../../../src/shared/domain/constants.js';
 import { expect } from '../../../test-helper.js';
 import { databaseBuilder } from '../../../tooling/databases.js';
 import { buildLearningContent as learningContentBuilder } from '../../../tooling/learning-content-builder/index.js';
