+++ api/tests/devcomp/integration/application/modules-metadata/modules-metadata-route-test.js
@@ -1,5 +1,5 @@
-import { PIX_ADMIN } from '../../../../../src/authorization/domain/constants.js';
 import * as moduleUnderTest from '../../../../../src/devcomp/application/modules-metadata/module-metadata-route.js';
+import { PIX_ADMIN } from '../../../../../src/shared/domain/constants.js';
 import { expect } from '../../../../test-helper.js';
 import { databaseBuilder } from '../../../../tooling/databases.js';
 import { HttpTestServer } from '../../../../tooling/server/http-test-server.js';
