+++ api/tests/devcomp/integration/application/trainings/training-route_test.js
@@ -1,9 +1,9 @@
 import sinon from 'sinon';
 
-import { PIX_ADMIN } from '../../../../../src/authorization/domain/constants.js';
 import { trainingController } from '../../../../../src/devcomp/application/trainings/training-controller.js';
 import * as moduleUnderTest from '../../../../../src/devcomp/application/trainings/training-route.js';
 import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
+import { PIX_ADMIN } from '../../../../../src/shared/domain/constants.js';
 import { expect } from '../../../../test-helper.js';
 import { getAdminRoleStub } from '../../../../tooling/mocks/security-pre-handlers.mock.js';
 import { HttpTestServer } from '../../../../tooling/server/http-test-server.js';
