+++ api/tests/identity-access-management/acceptance/application/token.route.test.js
@@ -1,7 +1,7 @@
 import querystring from 'node:querystring';
 
 import { createServer } from '../../../../server.js';
-import { PIX_ADMIN } from '../../../../src/authorization/domain/constants.js';
+import { PIX_ADMIN } from '../../../../src/shared/domain/constants.js';
 import { expect } from '../../../test-helper.js';
 import { databaseBuilder, knex } from '../../../tooling/databases.js';
 import { generateInjectOptions } from '../../../tooling/test-utils/http-server.js';
