+++ api/tests/prescription/learner-management/acceptance/application/organization-import-route_test.js
@@ -1,5 +1,5 @@
 import { createServer } from '../../../../../server.js';
-import { PIX_ADMIN } from '../../../../../src/authorization/domain/constants.js';
+import { PIX_ADMIN } from '../../../../../src/shared/domain/constants.js';
 import { ORGANIZATION_FEATURE } from '../../../../../src/shared/domain/constants.js';
 import { CsvImportError } from '../../../../../src/shared/domain/errors.js';
 import { Membership } from '../../../../../src/shared/domain/models/Membership.js';
