+++ api/tests/prescription/learner-management/unit/application/organization-learners-controller_test.js
@@ -1,8 +1,8 @@
 import sinon from 'sinon';
 
-import { CLIENTS, PIX_ADMIN } from '../../../../../src/authorization/domain/constants.js';
 import { organizationLearnersController } from '../../../../../src/prescription/learner-management/application/organization-learners-controller.js';
 import { usecases } from '../../../../../src/prescription/learner-management/domain/usecases/index.js';
+import { CLIENTS, PIX_ADMIN } from '../../../../../src/shared/domain/constants.js';
 import { expect } from '../../../../test-helper.js';
 import { hFake } from '../../../../tooling/mocks/hapi.mock.js';
