+++ api/tests/quest/acceptance/application/combined-course-route_test.js
@@ -1,11 +1,11 @@
 import iconv from 'iconv-lite';
 
 import { createServer } from '../../../../server.js';
-import { PIX_ADMIN } from '../../../../src/authorization/domain/constants.js';
 import {
   OrganizationLearnerParticipationStatuses,
   OrganizationLearnerParticipationTypes,
 } from '../../../../src/quest/domain/models/OrganizationLearnerParticipation.js';
+import { PIX_ADMIN } from '../../../../src/shared/domain/constants.js';
 import { expect } from '../../../test-helper.js';
 import { databaseBuilder } from '../../../tooling/databases.js';
 import { generateAuthenticatedUserRequestHeaders } from '../../../tooling/test-utils/http-server.js';
