+++ api/tests/team/integration/domain/usecases/get-user-teams-info.usecase.test.js
@@ -1,4 +1,4 @@
-import { PIX_ADMIN } from '../../../../../src/authorization/domain/constants.js';
+import { PIX_ADMIN } from '../../../../../src/shared/domain/constants.js';
 import { usecases } from '../../../../../src/team/domain/usecases/index.js';
 import { expect } from '../../../../test-helper.js';
 import { databaseBuilder } from '../../../../tooling/databases.js';
