+++ api/tests/team/unit/domain/usecases/update-admin-member.usecase.test.js
@@ -1,6 +1,6 @@
 import sinon from 'sinon';
 
-import { PIX_ADMIN } from '../../../../../src/authorization/domain/constants.js';
+import { PIX_ADMIN } from '../../../../../src/shared/domain/constants.js';
 import { updateAdminMember } from '../../../../../src/team/domain/usecases/update-admin-member.usecase.js';
 import { expect } from '../../../../test-helper.js';
